Pull data-input apart and split out the csv parser.
